Moved sendAudio to a page in components
